+++ lib/features/Meals/presentation/widgets/add new meal sheet.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 
 import '../../../../generated/l10n.dart';
-import '../../../../global/globalMethods.dart';
-import '../../../../global/widgets/defaultButton.dart';
-import '../../../../global/widgets/defaultTextFormField.dart';
+import '../../../../global/global_methods.dart';
+import '../../../../global/widgets/default_button.dart';
+import '../../../../global/widgets/default_text_form_field.dart';
 import '../../domain/entities/meal.dart';
 import '../bloc/MealsCubit/MealsCubit.dart';
 
@@ -80,10 +80,10 @@ void addNewMealSheet(context) {
                               if (formKey.currentState!.validate()) {
                                 MealsCubit.get(context).addMeal(Meal(
                                   protein:
-                                  int.parse(proteinController.text),
-                                  fat: int.parse(fatController.text),
-                                  carb: int.parse(carbController.text),
-                                  calories: int.parse(caloriesController.text),
+                                  double.parse(proteinController.text),
+                                  fat: double.parse(fatController.text),
+                                  carb: double.parse(carbController.text),
+                                  calories: double.parse(caloriesController.text),
                                   name: nameController.text,
                                 ));
                                 nameController.clear();
